feat: S3 scan object module
Add a module that creates a Lambda function that triggrs a Scan Files
ClamAV scan of new S3 objects created in a given upload bucket.

Once the scan is complete, the results are then published to an
SNS topic, which updates the scanned S3 object with a tag showing
the scan verdict.