fix: force destroy Athena DB and workgroup
Update the Athena access log module to set force_destroy on
the database and workgroup.  This will make it easier for
projects using the module to remove these resources.

Force destroy is safe in these cases because the database is
only a copy of the data which exists in the S3
access log bucket.