feat: add client VPN module with federated auth
Add a module that creates a client VPN in a given VPC.  The
module authorizes, and optionally associates, the VPN with
a set of subnets in the VPC.

Authentication and authorization is handled by SAML app and
groups managed in IAM Identity Cener.

By default, the VPN is setup to use `udp` and split tunnelling
with the VPC's internal DNS resolver.  This allows connected
users to continue to use the internet.