feat: add variable number of azs to VPC
removes the high_availability flag and replaces it with a variable that
accepts a number of azs to use for the vpc.

this will fail at compile time if the number of azs is greater than the
number of azs in the region.

removes deprecated vpc = true flag from the module and replaces it with
domain = "vpc" to match aws provider v5 syntax.

BREAKING CHANGE - will no longer work with provders < 5