fix: ECS module when not creating cluster (#396)
Update the ECS module so that it works as expected when
it does not handle the cluster creation.

Update the module to use Terraform unit tests rather than
Terratest.  For now these are just replicating the Terratests
and performing a smoke test to ensure the module can
successfully apply.  Assertions will be added in a future
PR 😅