fix: RDS activity stream lambda external provider
Update to use an `external` provider to create the decrypt Lambda layer
dependency zip file.  This allows us to create a deterministic zip file that
will only change when the `requirements.txt` file is updated.

As a result, the Lambda layer will only be updated when there are new
dependencies or version updates.