feat: create an RDS cluster without a proxy
Update the RDS module so that it can be used to create
RDS clusters without an associated proxy.  The change has includes
a migration file so that existing users of the module will be able
to upgrade without any breaking changes.

This is being done for projects that handle the connection pooling
internally and do not want to pay extra for the proxy.