feat: automated WAF IP blocklist
Add a module that creates a WAF IP set that contains IPv4 addresses
that have crossed a threshold of blocked requests within the past 24 hours.

The blocklist is managed by a Lambda function that runs on a schedule
and queries the WAF logs to determine which IP addresses should be
added.