fix: simple static website when using existing hosted zone
Update the module so that is works with an existing hosted zone when
the hosted zone and the module are being created by the same Terraform
apply operation.

This will be a breaking change since now existing users of the module with
a custom hosted zone or ACM certificate will need to specify the new
boolean inputs.